pyproject.toml
[project]
name = "fathom"
version = "0.1.0"
description = "AI-powered web research agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "aiosqlite",
    "gitpython",
    "crawl4ai",
    "boto3",
    "weasyprint",
    "python-multipart",
    "markdown",
    "python-dotenv",
    "openai>=1.0.0",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = []
